+++ demo/symfony8/config/bundles.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 use Nowo\HotReloadBundle\NowoHotReloadBundle;
 use Nowo\SlideToConfirmBundle\NowoSlideToConfirmBundle;
 use Nowo\TwigInspectorBundle\NowoTwigInspectorBundle;
+use Pentatrion\ViteBundle\PentatrionViteBundle;
 use Symfony\Bundle\DebugBundle\DebugBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\TwigBundle\TwigBundle;
@@ -14,6 +15,7 @@ use Twig\Extra\TwigExtraBundle\TwigExtraBundle;
 return [
     FrameworkBundle::class          => ['all' => true],
     TwigBundle::class               => ['all' => true],
+    PentatrionViteBundle::class     => ['all' => true],
     DebugBundle::class              => ['dev' => true, 'test' => true],
     WebProfilerBundle::class        => ['dev' => true, 'test' => true],
     StimulusBundle::class           => ['all' => true],